Move wrapper into it's own file so it can be re-used
